Moved lookup.js to parts, styles to const in styles.js, added escape keypress logic -kb
